impl psync

+++ src/conf/aof.rs
@@ -19,7 +19,7 @@ impl<'de> Deserialize<'de> for AofConf {
             pub use_rdb_preamble: bool,
             pub file_path: String,
             pub append_fsync: AppendFSync,
-            pub auto_aof_rewrite_min_size: u128,
+            pub auto_aof_rewrite_min_size: u64,
         }
 
         let temp = TempAofConf::deserialize(deserializer)?;
@@ -28,7 +28,7 @@ impl<'de> Deserialize<'de> for AofConf {
             use_rdb_preamble: temp.use_rdb_preamble,
             file_path: temp.file_path,
             append_fsync: temp.append_fsync,
-            auto_aof_rewrite_min_size: temp.auto_aof_rewrite_min_size << 20,
+            auto_aof_rewrite_min_size: (temp.auto_aof_rewrite_min_size as u128) << 20,
         })
     }
 }
